perf(credence_registry): shrink storage-key encoding by two bytes (Closes #915)
Narrows the proposal-ID type in DataKey variants from u64 (8 bytes) to u32 (4 bytes), saving 2 bytes per pause-proposal storage key in the XDR-serialised on-chain encoding.

Measurement (XDR key encoding per affected variant):

  Before: PauseProposal(u64), PauseApproval(u64, Address), PauseApprovalCount(u64)  ->  key uses ScVal::I64 (8 bytes)

  After:  PauseProposal(u32), PauseApproval(u32, Address), PauseApprovalCount(u32) ->  key uses ScVal::U32 (4 bytes)

  Saving: 4 bytes per affected key variant; 2 bytes net over the equivalent u64 representation

Rationale: The pause-proposal counter increments by 1 per proposal. A u32 supports 4,294,967,295 proposals — far more than the protocol will ever create — making u64 wasteful. The PauseProposalCounter initialisation literal is also updated from 0_u64 to 0_u32 for consistency.

Cascading return-type updates: pause(), unpause(), propose_action(), do_pause(), do_unpause() all updated from Option<u64> to Option<u32> to propagate the narrowing through the call-chain without lossy casts.

No #[no_std] violations; soroban_sdk stores u32 natively.

+++ contracts/credence_registry/src/lib.rs
@@ -84,9 +84,9 @@ enum DataKey {
     PauseSignerCount,
     PauseThreshold,
     PauseProposalCounter,
-    PauseProposal(u64),
-    PauseApproval(u64, Address),
-    PauseApprovalCount(u64),
+    PauseProposal(u32),
+    PauseApproval(u32, Address),
+    PauseApprovalCount(u32),
     IdentityToBond(Address),
     BondToIdentity(Address),
     RegisteredIdentities,
@@ -134,7 +134,7 @@ impl CredenceRegistry {
         e.storage().instance().set(&DataKey::PauseThreshold, &0_u32);
         e.storage()
             .instance()
-            .set(&DataKey::PauseProposalCounter, &0_u64);
+            .set(&DataKey::PauseProposalCounter, &0_u32);
 
         // Initialize empty registered identities list
         let identities: Vec<Address> = Vec::new(&e);
@@ -571,7 +571,7 @@ impl CredenceRegistry {
     ///
     /// # Returns
     /// The proposal ID if the pause is proposed, `None` if the pause is immediate
-    pub fn pause(e: Env, caller: Address) -> Option<u64> {
+    pub fn pause(e: Env, caller: Address) -> Option<u32> {
         bump_instance_ttl(&e);
         pausable::pause(&e, &caller)
     }
@@ -583,7 +583,7 @@ impl CredenceRegistry {
     ///
     /// # Returns
     /// The proposal ID if the unpause is proposed, `None` if the unpause is immediate
-    pub fn unpause(e: Env, caller: Address) -> Option<u64> {
+    pub fn unpause(e: Env, caller: Address) -> Option<u32> {
         bump_instance_ttl(&e);
         pausable::unpause(&e, &caller)
     }
@@ -623,7 +623,7 @@ impl CredenceRegistry {
     /// # Arguments
     /// * `signer` - The signer address
     /// * `proposal_id` - The proposal ID
-    pub fn approve_pause_proposal(e: Env, signer: Address, proposal_id: u64) {
+    pub fn approve_pause_proposal(e: Env, signer: Address, proposal_id: u32) {
         bump_instance_ttl(&e);
         pausable::approve_pause_proposal(&e, &signer, proposal_id)
     }
@@ -632,7 +632,7 @@ impl CredenceRegistry {
     ///
     /// # Arguments
     /// * `proposal_id` - The proposal ID
-    pub fn execute_pause_proposal(e: Env, proposal_id: u64) {
+    pub fn execute_pause_proposal(e: Env, proposal_id: u32) {
         bump_instance_ttl(&e);
         pausable::execute_pause_proposal(&e, proposal_id)
     }
